Lot's of stuff. Mostly palette and load list related.
Palettes seem to be organized in 16 16-color sets, just like in the SNES.
There are palette fading functions that accept different fading values
for each color, which, in addition to fade-ins/fade-outs, are used to dim
the screen, tint it, etc. (For example, after bombing the ship computer in
the second level.) There are actually two of those color values, and they
are ORed together. I don't know what that tries to achieve, nor what the
second color is used for. There are also two palettes, The source palette,
stored in palette1, and the destination in palette2, though their roles may
not be fixed on all uses.

The game seems to use what I called "load lists" to determine what
resources to load. Starting from a default load list id (0x27), that id is
used to index into a table of chunk ids, and that chunk is then loaded,
containing the list contents, as well as several other control variables,
one of them being the next list id (so it appears that they are chained that
way). The list (actually some sort of stream) is then interpreted by various
functions that load the resources it indicates.