Added more VGA emulation, and video output.
So as to not to have to worry about making sure I get the way the game uses
the VGA correctly, I'm going to be using an abstaction layer that will take
care of palettes, the planar format of video memory as well as registers
controlling scrolling and other misc. effects.  This should also integrate
nicely with scalers, if I decide to do so in the future.

Building now requires SDL 1.3.