Keyboard input handling.
Since the original game reads old PS/2 scancodes, I had to create a
table that will remap the SDL scancodes, which are based on the USB
standard to those old scancodes. Time will tell if that was a wise
decision or a huge waste of time.

After that tranlation is done, the game consults additional tables to
know what game buttons correspond to each key, and also to do a partial
key -> ASCII translation, used for the password screen.

The game buttons are stored as a bitfield, with one bit for each unique
gameplay action.