Initial implementation of object script VM.
Objects are scripted using a (quite powerful and non-sandboxed...)
variable-lenght-instruction bytecode that is interpreted by the game.
There are at least ~250 opcodes, all of which need to be individually
implemented.

The VM is implemented as a simple loop with a bit of state that fetches
opcodes from the instruction stream and dispatches on an array.