Additions and fixes to the script vm.
Implemented a bunch of new opcodes as well as fixed bugs in existing
ones. This makes the Interplay logo work (can be waited/skipped). The
(seemingly identical) Silicon & Synapse logo still deadlocks though, I'm
probably neglecting to reset some value, in one of the stubbed
functions.

As mentioned before, the VM isn't very sandboxed. In fact, the script
files are very machine/program dependent. Several opcodes involve
allowing the script to arbitrarily read from program memory, using
addresses in the script. This of course won't work since the memory
layout is incompatible with the DOS original. My solution is hope that
the accesses are limited to a few specific addresses and manually
translate them.

There's some also a subtle effect where the script can return values to
the VM and indicate which object will be the next one to be processed
which I hadn't noticed before, but now implemented.

Opcode 0x2F spawns a (more limited) sub VM with different opcodes. I'm
not sure what exactly it's purpose it yet.

In addition to all that I've also fixed some byte/word indexing issues
in vm_checkBit.