Implemented correct timing.
The original game used a cute contraption with the system timer that
synced it to the VBlank interval of the display, ensuring accurate
frame timing relative to the display. (Basically, a way to get a VBlank
interrupt, where the PC doesn't support one.)

This essentially means that timer_wait has to count VBlanks when
waiting. I use actual VSyncing instead, if available, or SDL_Delay with
a bit of loop spinning as a fallback. Regardless, I keep track of frame
times and missed frames using the high resolution timer to ensure that
missed frames are dropped, ensuring time smooth movement, instead of
displaying all frames but having jitter. I think this is the best
solution, but seeing it in practice later will be the only way to be
sure.

While examining how the video code interacted with timing, I also
realized that the VGA only reads the screen start registers at the
beginning of a new frame. That means that vga_setStartAddress should
*not* call vga_present. Since the wait timing is directly coupled with
VSync I update the emulated VGA frame there instead. This clears up why
the original didn't bother waiting for VSync to update the screen
pointer. Setting the palettes or pixel scrolling takes immediate effect
as far as I know, but the game doesn't rely on that and it isn't worth
emulating.